requirements.txt
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.3
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.2.1
pytest-cov==5.0.0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.2
urllib3==2.2.1
